vivbliss_scraper/requirements.txt
scrapy>=2.11.0
pymongo>=4.6.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
mongomock>=4.1.2
scrapy-fake-useragent>=1.4.4
python-dotenv>=1.0.0
pyyaml>=6.0.0